+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # v5.0.0
 
+## Features
+- 将模块作为标准 npm package 彻底与 nodebot 分离。
+- 增加了可外部调用的 HTTP API
+- 大幅扩展了 Message 对象，包括 static 方法以及新属性。
+- 将配置文件统一放置到 /config 目录下
+- 现在在调用 init 脚本时也会接受到空 Message 对象
+
+## Bugs
+- 修复了命令解析的重大 bug
+- 修复了其他各处的 bugs
 
 ---